+++ lib/src/date/days_picker.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 
 import '../shared/device_orientation_builder.dart';
+import '../shared/types.dart';
 import '../shared/utils.dart';
 import 'days_view.dart';
 import '../shared/header.dart';
@@ -72,13 +73,13 @@ class DaysPicker extends StatefulWidget {
     this.centerLeadingDate = false,
     this.previousPageSemanticLabel = 'Previous Day',
     this.nextPageSemanticLabel = 'Next Day',
+    this.disabledDayPredicate,
   }) {
     assert(!minDate.isAfter(maxDate), "minDate can't be after maxDate");
     assert(
       () {
         if (initialDate == null) return true;
-        final init =
-            DateTime(initialDate!.year, initialDate!.month, initialDate!.day);
+        final init = DateTime(initialDate!.year, initialDate!.month, initialDate!.day);
 
         final min = DateTime(minDate.year, minDate.month, minDate.day);
 
@@ -89,8 +90,7 @@ class DaysPicker extends StatefulWidget {
     assert(
       () {
         if (initialDate == null) return true;
-        final init =
-            DateTime(initialDate!.year, initialDate!.month, initialDate!.day);
+        final init = DateTime(initialDate!.year, initialDate!.month, initialDate!.day);
 
         final max = DateTime(maxDate.year, maxDate.month, maxDate.day);
         return init.isBefore(max) || init.isAtSameMomentAs(max);
@@ -234,6 +234,9 @@ class DaysPicker extends StatefulWidget {
   /// Semantic label for button to go to the next page.
   final String? nextPageSemanticLabel;
 
+  /// A predicate function used to determine if a given day should be disabled.
+  final DatePredicate? disabledDayPredicate;
+
   @override
   State<DaysPicker> createState() => _DaysPickerState();
 }
@@ -248,13 +251,10 @@ class _DaysPickerState extends State<DaysPicker> {
 
   @override
   void initState() {
-    final clampedInitailDate = DateUtilsX.clampDateToRange(
-        max: widget.maxDate, min: widget.minDate, date: DateTime.now());
-    _displayedMonth =
-        DateUtils.dateOnly(widget.initialDate ?? clampedInitailDate);
-    _selectedDate = widget.selectedDate != null
-        ? DateUtils.dateOnly(widget.selectedDate!)
-        : null;
+    final clampedInitailDate =
+        DateUtilsX.clampDateToRange(max: widget.maxDate, min: widget.minDate, date: DateTime.now());
+    _displayedMonth = DateUtils.dateOnly(widget.initialDate ?? clampedInitailDate);
+    _selectedDate = widget.selectedDate != null ? DateUtils.dateOnly(widget.selectedDate!) : null;
     _pageController = PageController(
       initialPage: DateUtils.monthDelta(widget.minDate, _displayedMonth!),
     );
@@ -268,10 +268,9 @@ class _DaysPickerState extends State<DaysPicker> {
     // but for makeing debuging easy, we will navigate to the initial date again
     // if it changes.
     if (oldWidget.initialDate != widget.initialDate) {
-      final clampedInitailDate = DateUtilsX.clampDateToRange(
-          max: widget.maxDate, min: widget.minDate, date: DateTime.now());
-      _displayedMonth =
-          DateUtils.dateOnly(widget.initialDate ?? clampedInitailDate);
+      final clampedInitailDate =
+          DateUtilsX.clampDateToRange(max: widget.maxDate, min: widget.minDate, date: DateTime.now());
+      _displayedMonth = DateUtils.dateOnly(widget.initialDate ?? clampedInitailDate);
 
       _pageController.jumpToPage(
         DateUtils.monthDelta(widget.minDate, _displayedMonth!),
@@ -279,9 +278,7 @@ class _DaysPickerState extends State<DaysPicker> {
     }
 
     if (oldWidget.selectedDate != widget.selectedDate) {
-      _selectedDate = widget.selectedDate != null
-          ? DateUtils.dateOnly(widget.selectedDate!)
-          : null;
+      _selectedDate = widget.selectedDate != null ? DateUtils.dateOnly(widget.selectedDate!) : null;
     }
     super.didUpdateWidget(oldWidget);
   }
@@ -332,8 +329,7 @@ class _DaysPickerState extends State<DaysPicker> {
           color: colorScheme.onSurface.withOpacity(0.30),
         );
 
-    final BoxDecoration disbaledCellsDecoration =
-        widget.disabledCellsDecoration;
+    final BoxDecoration disbaledCellsDecoration = widget.disabledCellsDecoration;
 
     //
     //! current
@@ -363,12 +359,11 @@ class _DaysPickerState extends State<DaysPicker> {
           color: colorScheme.onPrimary,
         );
 
-    final BoxDecoration selectedCellDecoration =
-        widget.selectedCellDecoration ??
-            BoxDecoration(
-              color: colorScheme.primary,
-              shape: BoxShape.circle,
-            );
+    final BoxDecoration selectedCellDecoration = widget.selectedCellDecoration ??
+        BoxDecoration(
+          color: colorScheme.primary,
+          shape: BoxShape.circle,
+        );
 
     //
     //
@@ -387,13 +382,11 @@ class _DaysPickerState extends State<DaysPicker> {
 
     //
     //! splash
-    final splashColor = widget.splashColor ??
-        selectedCellDecoration.color?.withOpacity(0.3) ??
-        colorScheme.primary.withOpacity(0.3);
+    final splashColor =
+        widget.splashColor ?? selectedCellDecoration.color?.withOpacity(0.3) ?? colorScheme.primary.withOpacity(0.3);
 
-    final highlightColor = widget.highlightColor ??
-        selectedCellDecoration.color?.withOpacity(0.3) ??
-        colorScheme.primary.withOpacity(0.3);
+    final highlightColor =
+        widget.highlightColor ?? selectedCellDecoration.color?.withOpacity(0.3) ?? colorScheme.primary.withOpacity(0.3);
     //
     //
 
@@ -454,24 +447,20 @@ class _DaysPickerState extends State<DaysPicker> {
                 scrollDirection: Axis.horizontal,
                 key: _pageViewKey,
                 controller: _pageController,
-                itemCount:
-                    DateUtils.monthDelta(widget.minDate, widget.maxDate) + 1,
+                itemCount: DateUtils.monthDelta(widget.minDate, widget.maxDate) + 1,
                 onPageChanged: (monthPage) {
-                  final DateTime monthDate =
-                      DateUtils.addMonthsToMonthDate(widget.minDate, monthPage);
+                  final DateTime monthDate = DateUtils.addMonthsToMonthDate(widget.minDate, monthPage);
 
                   setState(() {
                     _displayedMonth = monthDate;
                   });
                 },
                 itemBuilder: (context, index) {
-                  final DateTime month =
-                      DateUtils.addMonthsToMonthDate(widget.minDate, index);
+                  final DateTime month = DateUtils.addMonthsToMonthDate(widget.minDate, index);
 
                   return DaysView(
                     key: ValueKey<DateTime>(month),
-                    currentDate: DateUtils.dateOnly(
-                        widget.currentDate ?? DateTime.now()),
+                    currentDate: DateUtils.dateOnly(widget.currentDate ?? DateTime.now()),
                     maxDate: DateUtils.dateOnly(widget.maxDate),
                     minDate: DateUtils.dateOnly(widget.minDate),
                     displayedMonth: month,
@@ -488,6 +477,7 @@ class _DaysPickerState extends State<DaysPicker> {
                     highlightColor: highlightColor,
                     splashColor: splashColor,
                     splashRadius: widget.splashRadius,
+                    disabledDayPredicate: widget.disabledDayPredicate,
                     onChanged: (value) {
                       setState(() {
                         _selectedDate = value;