+++ test/date/days_view_test.dart
@@ -6,8 +6,7 @@ import 'package:flutter_localizations/flutter_localizations.dart';
 
 void main() {
   group('DaysView', () {
-    testWidgets('should have no selected day when selectedDate is null',
-        (WidgetTester tester) async {
+    testWidgets('should have no selected day when selectedDate is null', (WidgetTester tester) async {
       final DateTime currentDate = DateTime.now();
 
       await tester.pumpWidget(
@@ -16,10 +15,8 @@ void main() {
             child: DaysView(
               currentDate: currentDate,
               onChanged: (DateTime date) {},
-              minDate: DateTime(
-                  currentDate.year - 2, currentDate.month, currentDate.day),
-              maxDate: DateTime(
-                  currentDate.year + 2, currentDate.month, currentDate.day),
+              minDate: DateTime(currentDate.year - 2, currentDate.month, currentDate.day),
+              maxDate: DateTime(currentDate.year + 2, currentDate.month, currentDate.day),
               displayedMonth: currentDate,
               currentDateTextStyle: const TextStyle(),
               daysOfTheWeekTextStyle: const TextStyle(),
@@ -40,8 +37,7 @@ void main() {
       final Finder selectedDayFinder = find.byWidgetPredicate((widget) {
         if (widget is Container && widget.decoration != null) {
           final BoxDecoration decoration = widget.decoration as BoxDecoration;
-          return decoration.border == null &&
-              decoration.shape == BoxShape.circle;
+          return decoration.border == null && decoration.shape == BoxShape.circle;
         }
         return false;
       });
@@ -49,8 +45,7 @@ void main() {
       expect(selectedDayFinder, findsNothing);
     });
 
-    testWidgets('today should be the only cell that highlighted with border.',
-        (WidgetTester tester) async {
+    testWidgets('today should be the only cell that highlighted with border.', (WidgetTester tester) async {
       final DateTime currentDate = DateTime.now();
 
       await tester.pumpWidget(
@@ -59,18 +54,15 @@ void main() {
             child: DaysView(
               currentDate: currentDate,
               onChanged: (DateTime date) {},
-              minDate: DateTime(
-                  currentDate.year - 2, currentDate.month, currentDate.day),
-              maxDate: DateTime(
-                  currentDate.year + 2, currentDate.month, currentDate.day),
+              minDate: DateTime(currentDate.year - 2, currentDate.month, currentDate.day),
+              maxDate: DateTime(currentDate.year + 2, currentDate.month, currentDate.day),
               displayedMonth: currentDate,
               currentDateTextStyle: const TextStyle(),
               daysOfTheWeekTextStyle: const TextStyle(),
               enabledCellsTextStyle: const TextStyle(),
               selectedDayTextStyle: const TextStyle(),
               disabledCellsTextStyle: const TextStyle(),
-              currentDateDecoration:
-                  BoxDecoration(shape: BoxShape.circle, border: Border.all()),
+              currentDateDecoration: BoxDecoration(shape: BoxShape.circle, border: Border.all()),
               enabledCellsDecoration: const BoxDecoration(),
               selectedDayDecoration: const BoxDecoration(),
               disabledCellsDecoration: const BoxDecoration(),
@@ -84,8 +76,7 @@ void main() {
       final Finder todayFinder = find.byWidgetPredicate((widget) {
         if (widget is Container && widget.decoration != null) {
           final BoxDecoration decoration = widget.decoration as BoxDecoration;
-          return decoration.border != null &&
-              decoration.shape == BoxShape.circle;
+          return decoration.border != null && decoration.shape == BoxShape.circle;
         }
         return false;
       });
@@ -93,8 +84,7 @@ void main() {
       expect(todayFinder, findsOneWidget);
     });
 
-    testWidgets(
-        'should be two widget highlighted, Today with border, and selected day with fill color.',
+    testWidgets('should be two widget highlighted, Today with border, and selected day with fill color.',
         (WidgetTester tester) async {
       final DateTime currentDate = DateTime.now();
       final DateTime selectedDate = DateTime(
@@ -110,10 +100,8 @@ void main() {
             child: DaysView(
               currentDate: currentDate,
               onChanged: (DateTime date) {},
-              minDate: DateTime(
-                  currentDate.year - 2, currentDate.month, currentDate.day),
-              maxDate: DateTime(
-                  currentDate.year + 2, currentDate.month, currentDate.day),
+              minDate: DateTime(currentDate.year - 2, currentDate.month, currentDate.day),
+              maxDate: DateTime(currentDate.year + 2, currentDate.month, currentDate.day),
               displayedMonth: currentDate,
               selectedDate: selectedDate,
               currentDateTextStyle: const TextStyle(),
@@ -121,10 +109,8 @@ void main() {
               enabledCellsTextStyle: const TextStyle(),
               selectedDayTextStyle: const TextStyle(),
               disabledCellsTextStyle: const TextStyle(),
-              currentDateDecoration:
-                  BoxDecoration(shape: BoxShape.circle, border: Border.all()),
-              selectedDayDecoration:
-                  const BoxDecoration(shape: BoxShape.circle),
+              currentDateDecoration: BoxDecoration(shape: BoxShape.circle, border: Border.all()),
+              selectedDayDecoration: const BoxDecoration(shape: BoxShape.circle),
               enabledCellsDecoration: const BoxDecoration(),
               disabledCellsDecoration: const BoxDecoration(),
               splashColor: Colors.black,
@@ -137,8 +123,7 @@ void main() {
       final Finder selectedDayFinder = find.byWidgetPredicate((widget) {
         if (widget is Container && widget.decoration != null) {
           final BoxDecoration decoration = widget.decoration as BoxDecoration;
-          return decoration.border == null &&
-              decoration.shape == BoxShape.circle;
+          return decoration.border == null && decoration.shape == BoxShape.circle;
         }
         return false;
       });
@@ -146,8 +131,7 @@ void main() {
       final Finder todayFinder = find.byWidgetPredicate((widget) {
         if (widget is Container && widget.decoration != null) {
           final BoxDecoration decoration = widget.decoration as BoxDecoration;
-          return decoration.border != null &&
-              decoration.shape == BoxShape.circle;
+          return decoration.border != null && decoration.shape == BoxShape.circle;
         }
         return false;
       });
@@ -156,8 +140,7 @@ void main() {
       expect(todayFinder, findsOneWidget);
     });
 
-    testWidgets(
-        'should be one widget highlighted, when selected day is not in the month displayed.',
+    testWidgets('should be one widget highlighted, when selected day is not in the month displayed.',
         (WidgetTester tester) async {
       final DateTime currentDate = DateTime.now();
       final DateTime selectedDate = DateTime(
@@ -173,10 +156,8 @@ void main() {
             child: DaysView(
               currentDate: currentDate,
               onChanged: (DateTime date) {},
-              minDate: DateTime(
-                  currentDate.year - 2, currentDate.month, currentDate.day),
-              maxDate: DateTime(
-                  currentDate.year + 2, currentDate.month, currentDate.day),
+              minDate: DateTime(currentDate.year - 2, currentDate.month, currentDate.day),
+              maxDate: DateTime(currentDate.year + 2, currentDate.month, currentDate.day),
               displayedMonth: currentDate,
               selectedDate: selectedDate,
               currentDateTextStyle: const TextStyle(),
@@ -184,11 +165,9 @@ void main() {
               enabledCellsTextStyle: const TextStyle(),
               selectedDayTextStyle: const TextStyle(),
               disabledCellsTextStyle: const TextStyle(),
-              currentDateDecoration:
-                  BoxDecoration(shape: BoxShape.circle, border: Border.all()),
+              currentDateDecoration: BoxDecoration(shape: BoxShape.circle, border: Border.all()),
               enabledCellsDecoration: const BoxDecoration(),
-              selectedDayDecoration:
-                  const BoxDecoration(shape: BoxShape.circle),
+              selectedDayDecoration: const BoxDecoration(shape: BoxShape.circle),
               disabledCellsDecoration: const BoxDecoration(),
               splashColor: Colors.black,
               highlightColor: Colors.black,
@@ -200,8 +179,7 @@ void main() {
       final Finder selectedDayFinder = find.byWidgetPredicate((widget) {
         if (widget is Container && widget.decoration != null) {
           final BoxDecoration decoration = widget.decoration as BoxDecoration;
-          return decoration.border == null &&
-              decoration.shape == BoxShape.circle;
+          return decoration.border == null && decoration.shape == BoxShape.circle;
         }
         return false;
       });
@@ -209,8 +187,7 @@ void main() {
       final Finder todayFinder = find.byWidgetPredicate((widget) {
         if (widget is Container && widget.decoration != null) {
           final BoxDecoration decoration = widget.decoration as BoxDecoration;
-          return decoration.border != null &&
-              decoration.shape == BoxShape.circle;
+          return decoration.border != null && decoration.shape == BoxShape.circle;
         }
         return false;
       });
@@ -218,13 +195,10 @@ void main() {
       expect(selectedDayFinder, findsNothing);
       expect(todayFinder, findsOneWidget);
     });
-    testWidgets('should throw assertion error if minDate > maxDate',
-        (WidgetTester tester) async {
+    testWidgets('should throw assertion error if minDate > maxDate', (WidgetTester tester) async {
       final DateTime currentDate = DateTime.now();
-      final DateTime maxDate =
-          DateTime.now().subtract(const Duration(days: 365 * 10));
-      final DateTime minDate =
-          DateTime.now().add(const Duration(days: 365 * 10));
+      final DateTime maxDate = DateTime.now().subtract(const Duration(days: 365 * 10));
+      final DateTime minDate = DateTime.now().add(const Duration(days: 365 * 10));
 
       expect(() async {
         await tester.pumpWidget(
@@ -254,8 +228,7 @@ void main() {
       }, throwsAssertionError);
     });
 
-    testWidgets('should disbale all the days before min date.',
-        (WidgetTester tester) async {
+    testWidgets('should disbale all the days before min date.', (WidgetTester tester) async {
       final DateTime currentDate = DateTime(2020, 1, 25);
       final DateTime minDate = DateTime(2020, 1, 10);
       final DateTime maxDate = DateTime(2020, 1, 30);
@@ -288,9 +261,7 @@ void main() {
       );
 
       final disabledDayFinder = find.byWidgetPredicate((widget) {
-        if (widget is ExcludeSemantics &&
-            widget.child is Container &&
-            (widget.child as Container).child is Center) {
+        if (widget is ExcludeSemantics && widget.child is Container && (widget.child as Container).child is Center) {
           final container = widget.child as Container;
           return (container.decoration as BoxDecoration).color == Colors.green;
         }
@@ -299,8 +270,7 @@ void main() {
       expect(disabledDayFinder, findsNWidgets(10));
     });
 
-    testWidgets('should disbale all the days after max date.',
-        (WidgetTester tester) async {
+    testWidgets('should disbale all the days after max date.', (WidgetTester tester) async {
       final DateTime currentDate = DateTime(2020, 1, 15);
       final DateTime minDate = DateTime(2020, 1, 1);
       final DateTime maxDate = DateTime(2020, 1, 20);
@@ -333,9 +303,7 @@ void main() {
       );
 
       final disabledDayFinder = find.byWidgetPredicate((widget) {
-        if (widget is ExcludeSemantics &&
-            widget.child is Container &&
-            (widget.child as Container).child is Center) {
+        if (widget is ExcludeSemantics && widget.child is Container && (widget.child as Container).child is Center) {
           final container = widget.child as Container;
           return (container.decoration as BoxDecoration).color == Colors.green;
         }
@@ -344,9 +312,53 @@ void main() {
       expect(disabledDayFinder, findsNWidgets(11));
     });
 
-    testWidgets(
-        'should show the correct first day of the week based on locale.',
-        (WidgetTester tester) async {
+    testWidgets('should disbale all the days provided in disabledDayPredicate.', (WidgetTester tester) async {
+      final DateTime currentDate = DateTime(2021);
+      final DateTime displayedMonth = DateTime(2023);
+      final DateTime minDate = DateTime(2020);
+      final DateTime maxDate = DateTime(2024);
+
+      await tester.pumpWidget(
+        MaterialApp(
+          home: Material(
+            child: DaysView(
+              currentDate: currentDate,
+              onChanged: (DateTime date) {},
+              disabledDayPredicate: (date) {
+                return date.weekday == DateTime.sunday || date.weekday == DateTime.saturday;
+              },
+              minDate: minDate,
+              maxDate: maxDate,
+              displayedMonth: displayedMonth,
+              currentDateTextStyle: const TextStyle(),
+              daysOfTheWeekTextStyle: const TextStyle(),
+              enabledCellsTextStyle: const TextStyle(),
+              selectedDayTextStyle: const TextStyle(),
+              disabledCellsTextStyle: const TextStyle(),
+              currentDateDecoration: const BoxDecoration(),
+              enabledCellsDecoration: const BoxDecoration(),
+              selectedDayDecoration: const BoxDecoration(),
+              splashColor: Colors.black,
+              highlightColor: Colors.black,
+              disabledCellsDecoration: const BoxDecoration(
+                color: Colors.green,
+              ),
+            ),
+          ),
+        ),
+      );
+
+      final disabledDayFinder = find.byWidgetPredicate((widget) {
+        if (widget is ExcludeSemantics && widget.child is Container && (widget.child as Container).child is Center) {
+          final container = widget.child as Container;
+          return (container.decoration as BoxDecoration).color == Colors.green;
+        }
+        return false;
+      });
+      expect(disabledDayFinder, findsNWidgets(9));
+    });
+
+    testWidgets('should show the correct first day of the week based on locale.', (WidgetTester tester) async {
       const uSLocale = Locale('en', 'US');
 
       await GlobalMaterialLocalizations.delegate.load(uSLocale);
@@ -355,8 +367,7 @@ void main() {
       final DateTime minDate = DateTime(2020, 1, 1);
       final DateTime maxDate = DateTime(2020, 1, 20);
 
-      final List<String> weekdayNames =
-          intl.DateFormat('', 'en').dateSymbols.SHORTWEEKDAYS;
+      final List<String> weekdayNames = intl.DateFormat('', 'en').dateSymbols.SHORTWEEKDAYS;
 
       late final MaterialLocalizations localizations;
 
@@ -399,27 +410,22 @@ void main() {
       );
 
       final int firstDayOfWeekIndex = localizations.firstDayOfWeekIndex;
-      final String expectedFirstDayOfWeek =
-          weekdayNames[firstDayOfWeekIndex].toUpperCase();
+      final String expectedFirstDayOfWeek = weekdayNames[firstDayOfWeekIndex].toUpperCase();
 
       final disabledDayFinder = find.byWidgetPredicate((widget) {
-        if (widget is Center &&
-            widget.child is Text &&
-            (widget.child as Text).data == expectedFirstDayOfWeek) {
+        if (widget is Center && widget.child is Text && (widget.child as Text).data == expectedFirstDayOfWeek) {
           return true;
         }
         return false;
       });
 
-      final RenderBox renderBox =
-          tester.renderObject<RenderBox>(disabledDayFinder);
+      final RenderBox renderBox = tester.renderObject<RenderBox>(disabledDayFinder);
       final Offset topLeft = renderBox.localToGlobal(Offset.zero);
 
       expect(topLeft, equals(Offset.zero));
     });
 
-    testWidgets('should display days\' names with the correct color',
-        (WidgetTester tester) async {
+    testWidgets('should display days\' names with the correct color', (WidgetTester tester) async {
       const Color customColor = Colors.blue; // Replace with your specific color
 
       await tester.pumpWidget(
@@ -456,20 +462,17 @@ void main() {
         return false;
       });
 
-      expect(dayNameFinder,
-          findsNWidgets(7)); // Assuming there are 7 days in a week
+      expect(dayNameFinder, findsNWidgets(7)); // Assuming there are 7 days in a week
 
       // Verify that all day names have the correct color
       await tester.ensureVisible(dayNameFinder.first);
-      expect(
-          tester.widget<Text>(dayNameFinder.first).style?.color, customColor);
+      expect(tester.widget<Text>(dayNameFinder.first).style?.color, customColor);
 
       await tester.ensureVisible(dayNameFinder.last);
       expect(tester.widget<Text>(dayNameFinder.last).style?.color, customColor);
     });
 
-    testWidgets('should display enabled days with the correct color',
-        (WidgetTester tester) async {
+    testWidgets('should display enabled days with the correct color', (WidgetTester tester) async {
       const Color customColor = Colors.green;
 
       await tester.pumpWidget(
@@ -514,8 +517,7 @@ void main() {
       await tester.ensureVisible(enabledDayFinder.last);
     });
 
-    testWidgets('should display disabled days with the correct color',
-        (WidgetTester tester) async {
+    testWidgets('should display disabled days with the correct color', (WidgetTester tester) async {
       const Color customColor = Colors.green;
 
       await tester.pumpWidget(
@@ -559,8 +561,7 @@ void main() {
       await tester.ensureVisible(disabledDayFinder.last);
     });
 
-    testWidgets('should display today with the correct color',
-        (WidgetTester tester) async {
+    testWidgets('should display today with the correct color', (WidgetTester tester) async {
       const Color customColor = Colors.green;
 
       await tester.pumpWidget(
@@ -596,11 +597,9 @@ void main() {
       final Finder enabledDayFinder = find.byWidgetPredicate((widget) {
         if (widget is Container &&
             widget.decoration != null &&
-            (widget.decoration as BoxDecoration).border ==
-                Border.all(color: customColor) &&
+            (widget.decoration as BoxDecoration).border == Border.all(color: customColor) &&
             (widget.child as Center).child is Text &&
-            ((widget.child as Center).child as Text).style?.color ==
-                customColor) {
+            ((widget.child as Center).child as Text).style?.color == customColor) {
           return true;
         }
         return false;
@@ -613,8 +612,7 @@ void main() {
       await tester.ensureVisible(enabledDayFinder.last);
     });
 
-    testWidgets('should display selected day with the correct color',
-        (WidgetTester tester) async {
+    testWidgets('should display selected day with the correct color', (WidgetTester tester) async {
       const Color textColor = Colors.green;
       const Color fillColor = Colors.red;
 
@@ -654,8 +652,7 @@ void main() {
             (widget.decoration as BoxDecoration).border == null &&
             (widget.decoration as BoxDecoration).color == fillColor &&
             (widget.child as Center).child is Text &&
-            ((widget.child as Center).child as Text).style?.color ==
-                textColor) {
+            ((widget.child as Center).child as Text).style?.color == textColor) {
           return true;
         }
         return false;
@@ -665,8 +662,7 @@ void main() {
 
       await tester.ensureVisible(enabledDayFinder.first);
     });
-    testWidgets('should select the right date when tap.',
-        (WidgetTester tester) async {
+    testWidgets('should select the right date when tap.', (WidgetTester tester) async {
       final dateToSelect = DateTime(2020, 1, 4);
       DateTime? selectedDate;
 
@@ -697,19 +693,14 @@ void main() {
         ),
       );
 
-      final clickbaleWidget =
-          find.byWidgetPredicate((widget) => widget is InkResponse);
+      final clickbaleWidget = find.byWidgetPredicate((widget) => widget is InkResponse);
 
-      expect(
-          clickbaleWidget,
-          findsNWidgets(
-              31)); // Assuming there are 31 days in the displayed month
+      expect(clickbaleWidget, findsNWidgets(31)); // Assuming there are 31 days in the displayed month
 
       final Finder enabledDayFinder = find.byWidgetPredicate((widget) {
         if (widget is Container &&
             (widget.child as Center).child is Text &&
-            ((widget.child as Center).child as Text).data ==
-                dateToSelect.day.toString()) {
+            ((widget.child as Center).child as Text).data == dateToSelect.day.toString()) {
           return true;
         }
         return false;